Add several methods and changes
1. Soften the AR using the distance of the previous camera pose.

There is no "key point" indicating the upper left corner.
So there was an ambiguous match between img_corner and wor_corner.

To maintain consistency in the matching, we use the minimum distance of translation and rotation compared to the previous frame.

2. Add a code that accepts a possible combination of ball colors.
There shouldn't be two yellow balls or two white balls.

3. Change "at operation" to "ptr operation" for speed.

4. Fix additional errors